requirements.txt
requests
pyyaml
termcolor
emoji
